Finish moving columnNodeSupport and friends to utils
