+++ sql/core/src/main/scala/org/apache/spark/sql/classic/conversions.scala
@@ -21,7 +21,8 @@ import scala.language.implicitConversions
 import org.apache.spark.annotation.DeveloperApi
 import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.expressions.{Expression, NamedExpression}
-import org.apache.spark.sql.internal.{ColumnNodeToExpressionConverter, ExpressionUtils}
+import org.apache.spark.sql.internal.ExpressionUtils
+import org.apache.spark.sql.utils.ColumnNodeToExpressionConverter
 
 /**
  * Conversions from sql interfaces to the Classic specific implementation.
