+++ sql/core/src/main/scala/org/apache/spark/sql/util/columnNodeSupport.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.internal
+package org.apache.spark.sql.util
 
 import UserDefinedFunctionUtils.toScalaUDF
